Rename registry to vdr
