+++ vdr/logging/log.go
