+++ vdr/logging/log_test.go
