+++ vdr/network/mock.go
